+++ agent/statemanager/state_manager.go
@@ -59,7 +59,9 @@ const (
 	//   d) Added task cgroup related fields ('CPU', 'Memory', 'MemoryCPULimitsEnabled') to 'api.Task'
 	// 9) Add 'ipToTask' map to state file
 	// 10) Add 'healthCheckType' field in 'api.Container'
-	// 11) Add 'PrivateDNSName' field to 'api.ENI'
+	// 11)
+	//  a) Add 'PrivateDNSName' field to 'api.ENI'
+	//  b)Remove `AppliedStatus` field form 'api.Container'
 	ECSDataVersion = 11
 
 	// ecsDataFile specifies the filename in the ECS_DATADIR